# openapi.yaml
openapi: 3.0.0
info:
  title: apiTienda
  version: '1.1'
  description: API para gestionar una tienda en línea, incluyendo funcionalidades para manejar usuarios, productos, categorías, pedidos y detalles de pedidos. Permite realizar operaciones CRUD en estas entidades, facilitando la gestión completa de un sistema de comercio electrónico.
servers:
  - url: http://localhost:8080
    description: Servidor local
paths:
...
  /users/{userId}/orders/{orderId}/details:
    post:
      summary: Añadir un detalle a un pedido
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: long
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDetailInDTO'
      responses:
        '201':
          description: Detalle de pedido creado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Usuario, pedido o producto no encontrado
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/UserNotFound'
                  - $ref: '#/components/responses/OrderNotFound'
                  - $ref: '#/components/responses/ProductNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: Obtener detalles de un pedido
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: long
        - in: query
          name: discount
          schema:
            type: number
            format: float
        - in: query
          name: quantity
          schema:
            type: number
            format: float
        - in: query
          name: subtotal
          schema:
            type: number
            format: float
      responses:
        '200':
          description: Lista de detalles de pedido obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDetailOutDTO'
        '404':
          description: Usuario o pedido no encontrado
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/UserNotFound'
                  - $ref: '#/components/responses/OrderNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/orders/{orderId}/details/{detailId}:
    put:
      summary: Actualizar un detalle de pedido
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: detailId
          required: true
          schema:
            type: integer
            format: long
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDetailInDTO'
      responses:
        '200':
          description: Detalle de pedido actualizado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailOutDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Usuario, pedido, detalle de pedido o producto no encontrado
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/UserNotFound'
                  - $ref: '#/components/responses/OrderNotFound'
                  - $ref: '#/components/responses/OrderDetailNotFound'
                  - $ref: '#/components/responses/ProductNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      summary: Actualizar un detalle de pedido
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: detailId
          required: true
          schema:
            type: integer
            format: long
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDetailInDTO'
      responses:
        '200':
          description: Detalle de pedido actualizado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailOutDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Usuario, pedido, detalle de pedido o producto no encontrado
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/UserNotFound'
                  - $ref: '#/components/responses/OrderNotFound'
                  - $ref: '#/components/responses/OrderDetailNotFound'
                  - $ref: '#/components/responses/ProductNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Eliminar un detalle de pedido
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: detailId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Detalle de pedido eliminado con éxito
        '404':
          description: Usuario, pedido o detalle de pedido no encontrado
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/UserNotFound'
                  - $ref: '#/components/responses/OrderNotFound'
                  - $ref: '#/components/responses/OrderDetailNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
          format: long
        quantity:
          type: integer
        subtotal:
          type: number
          format: float
        discount:
          type: number
          format: float
        creationDate:
          type: string
          format: date
      example:
        id: 10001
        quantity: 2
        subtotal: 1199.98
        discount: 100.00
        creationDate: "2025-01-20"
    OrderDetailOutDTO:
      type: object
      properties:
        id:
          type: integer
          format: long
        orderId:
          type: integer
          format: long
        productId:
          type: integer
          format: long
        productName:
          type: string
        quantity:
          type: integer
        discount:
          type: number
          format: float
        subtotal:
          type: number
          format: float
      example:
        id: 10002
        orderId: 5001
        productId: 101
        productName: "Smartphone XYZ"
        quantity: 1
        discount: 50.00
        subtotal: 549.99
    OrderDetailInDTO:
      type: object
      required:
        - quantity
      properties:
        quantity:
          type: integer
        discount:
          type: number
          format: float
        productId:
          type: integer
          format: long
      example:
        quantity: 3
        discount: 75.00
        productId: 102
...
    OrderDetailNotFoundException:
      type: object
      properties:
        errorcode:
          type: integer
          example: 404
        message:
          type: string
          example: "OrderDetail con id: 123 no encontrado."
      example:
        errorcode: 404
        message: "OrderDetail con id: 123 no encontrado."
    OrderNotFoundException:
      type: object
      properties:
        errorcode:
          type: integer
          example: 404
        message:
          type: string
          example: "Pedido no encontrado"
      example:
        errorcode: 404
        message: "Pedido no encontrado"
    ProductNotFoundException:
      type: object
      properties:
        errorcode:
          type: integer
          example: 404
        message:
          type: string
          example: "Producto no encontrado"
      example:
        errorcode: 404
        message: "Producto no encontrado"
    UserNotFoundException:
      type: object
      properties:
        errorcode:
          type: integer
          example: 404
        message:
          type: string
          example: "El usuario no existe"
      example:
        errorcode: 404
        message: "El usuario no existe"
    ErrorResponse:
      type: object
      properties:
        errorcode:
          type: integer
        message:
          type: string
        errorMessages:
          type: object
          additionalProperties:
            type: string
      example:
        errorcode: 500
        message: "Internal Server Error"
  responses:
   BadRequest:
    description: Solicitud inválida
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/ErrorResponse'
        examples:
          BadRequest:
            value:
              errorcode: 400
              message: "Bad Request"
          ValidationError:
            value:
              errorcode: 400
              message: "Validation Error"
              errorMessages:
                name: "El nombre es requerido"
                email: "El formato del email es inválido"
...
   ProductNotFound:
    description: Producto no encontrado
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/ProductNotFoundException'
   UserNotFound:
    description: El usuario no existe
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/UserNotFoundException'
   OrderNotFound:
    description: Pedido no encontrado
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/OrderNotFoundException'
   OrderDetailNotFound:
    description: Detalle de pedido no encontrado
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/OrderDetailNotFoundException'
   InternalServerError:
    description: Error interno del servidor
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/ErrorResponse'